refactor: Thread 컴포넌트 합성 방식 변경
