+++ src/components/thread/index.tsx
@@ -1,1 +1,17 @@
-export * from './thread';
+import { Thread } from './thread';
+import { ThreadBody } from './thread-body';
+import { ThreadButtons } from './thread-buttons';
+import { ThreadContent } from './thread-content';
+import { ThreadExpandableContent } from './thread-expandable-content';
+import { ThreadFooter } from './thread-footer';
+import { ThreadHeader } from './thread-header';
+
+const Root = Thread;
+const Body = ThreadBody;
+const Content = ThreadContent;
+const ExpandableContent = ThreadExpandableContent;
+const Footer = ThreadFooter;
+const Header = ThreadHeader;
+const Buttons = ThreadButtons;
+
+export { Body, Buttons, Content, ExpandableContent, Footer, Header, Root };
