+++ src/components/thread/thread.tsx
@@ -1,9 +1,4 @@
 import { Flex } from '@/styled-system/jsx';
-import { ThreadBody } from './thread-body';
-import { ThreadContent } from './thread-content';
-import { ThreadExpandableContent } from './thread-expandable-content';
-import { ThreadFooter } from './thread-footer';
-import { ThreadHeader } from './thread-header';
 
 interface ThreadProps {
   children: React.ReactNode;
@@ -16,9 +11,3 @@ export function Thread({ children }: ThreadProps) {
     </Flex>
   );
 }
-
-Thread.Header = ThreadHeader;
-Thread.Body = ThreadBody;
-Thread.Footer = ThreadFooter;
-Thread.Content = ThreadContent;
-Thread.ExpandableContent = ThreadExpandableContent;
